Finally set a window icon on X11
Since the creation of minetest, it had no window icon on X11.
Now we have one.

Generating a header file that contains a giant integer array
is not really the nicest way to specify data, but the
alternatives require lots of code for this very simple task:

1. convert the png each time we start minetest. This would
	require lots of additional code, and probably be slower

2. store the image in a custom binary format. X11 doesn't
	specify a on-disk format, so we would have to invent
	one. Straight writing of the array to a file e.g. with
	a fwrite(buffer, sizeof(buffer), 1, stdout); call
	isn't enough as the format is not cross-platform due
	to endianness issues. Therefore we'd have to apply
	conversions on loading and storing. This would be
	lots of additional code as well.

In fact, the 8 bytes per long that are required for a
straight-forward binary format require *more* disk space
than if the icon is stored in ASCII format.